+++ packages/api/src/handlers/__tests__/projectSummaryHandler.test.ts
@@ -1,0 +1,4 @@
+
+describe("projectSummaryHandler", () => {
+  
+});
