+++ packages/api/src/handlers/projectSummaryHandler.ts
@@ -1,18 +1,20 @@
 import {
   ChainId,
-  QFVote,
 } from "../types";
 import {
-  fetchFromGraphQL,
   fetchRoundMetadata,
   getStrategyName,
-  handleResponse, summarizeQFVotes,
+  handleResponse,
 } from "../utils";
 import { Request, Response } from "express";
-import { BigNumber } from "ethers";
+import {
+  fetchQFContributionsForProject,
+  summarizeQFContributions
+} from "../votingStrategies/linearQuadraticFunding";
 
 /**
- * projectSummaryHandler is a function that handles HTTP requests for summary information for a given round and project.
+ * projectSummaryHandler is a function that handles HTTP requests
+ * for summary information for a given round and project.
  *
  * @param {Request} req - The incoming HTTP request.
  * @param {Response} res - The HTTP response that will be sent.
@@ -44,6 +46,7 @@ export const projectSummaryHandler = async (req: Request, res: Response) => {
  */
 export const getProjectsSummary = async (chainId: ChainId, roundId: string, projectIds: string[]): Promise<any> => {
   let results: any = [];
+
   // fetch metadata
   const metadata = await fetchRoundMetadata(chainId, roundId);
 
@@ -55,9 +58,9 @@ export const getProjectsSummary = async (chainId: ChainId, roundId: string, proj
   switch (strategyName) {
     case "LINEAR_QUADRATIC_FUNDING":
       // fetch votes
-      const votes = await fetchContributionsForProject(chainId, votingStrategyId, projectIds);
+      const contributions = await fetchQFContributionsForProject(chainId, votingStrategyId, projectIds);
       // fetch round stats
-      results =  await summarizeQFVotes(chainId, metadata, votes);
+      results =  await summarizeQFContributions(chainId, contributions);
       break;
     default:
       throw("error: unsupported voting strategy");
@@ -65,110 +68,3 @@ export const getProjectsSummary = async (chainId: ChainId, roundId: string, proj
 
   return results;
 }
-
-/**
- * fetchContributionsForProject is a function that fetches a list of contributions for a given project from a GraphQL API.
- *
- * @param {ChainId} chainId - The ID of the chain to fetch data from.
- * @param {string} votingStrategyId - The ID of the voting strategy to fetch data for.
- * @param {string[]} projectIds - An array of project IDs to filter the contributions by.
- * @returns {Promise<QFVote[]>} A promise that resolves to an array of QFVote objects.
- */
-export const fetchContributionsForProject = async (
-  chainId: ChainId,
-  votingStrategyId: string,
-  projectIds: string[],
-): Promise<QFVote[]> => {
-  let lastID: string = "";
-  const query = `
-    query GetContributionsForProject($votingStrategyId: String, $projectIds: [String]) {
-      votingStrategies(where:{
-        id: $votingStrategyId
-      }) {
-        votes(first: 1000, where: {
-          to_in: $projectIds
-        }) {
-          id
-          amount
-          token
-          from
-          to
-        }
-        round {
-          roundStartTime
-          roundEndTime
-          token
-        }
-
-      }
-
-    }
-  `;
-  const variables = { votingStrategyId, projectIds };
-  // fetch from graphql
-  const response = await fetchFromGraphQL(chainId, query, variables);
-
-  const votes: QFVote[] = [];
-
-  response.data?.votingStrategies[0]?.votes.map((vote: any) => {
-    votes.push({
-      amount: BigNumber.from(vote.amount),
-      token: vote.token,
-      contributor: vote.from,
-      projectId: vote.to, // TODO: update to projectID after contract upgrade
-    });
-    lastID = vote.id;
-  });
-
-  while (true) {
-    const query = `
-      query GetContributionsForProject($votingStrategyId: String, $lastID: String, $projectIds: [String]) {
-        votingStrategies(where:{
-          id: $votingStrategyId
-        }) {
-          votes(first: 1000, where: {
-              id_gt: $lastID
-              to_in: $projectIds
-          }) {
-            id
-            amount
-            token
-            from
-            to
-          }
-          round {
-            roundStartTime
-            roundEndTime
-            token
-          }
-        }
-      }
-    `;
-
-    // Fetch the next page of results from the GraphQL API
-    const response = await fetchFromGraphQL(chainId, query, {
-      votingStrategyId,
-      lastID,
-      projectIds,
-    });
-
-    // Check if the votes field is empty. If it is, stop paginating
-    if (response.data?.votingStrategies[0]?.votes.length === 0) {
-      break;
-    }
-
-    // Add the new votes to the list of votes
-    response.data?.votingStrategies[0]?.votes.map((vote: any) => {
-      votes.push({
-        amount: BigNumber.from(vote.amount),
-        token: vote.token,
-        contributor: vote.from,
-        projectId: vote.to, // TODO: update to projectID after contract upgrade
-      });
-      lastID = vote.id;
-    });
-  }
-
-  return votes;
-};
-
