+++ packages/api/src/utils.ts
@@ -5,10 +5,8 @@ import {
   ChainId,
   ChainName,
   RoundMetadata,
-  DenominationResponse, QFVote, QFVoteSummary,
+  DenominationResponse
 } from "./types";
-import {BigNumber} from "ethers";
-import {formatUnits} from "ethers/lib/utils";
 
 /**
  * Fetch subgraph network for provided web3 network
@@ -41,10 +39,10 @@ export const getGraphQLEndpoint = (chainId: ChainId) => {
 
 
 /**
- * Returns USDC address based on chain Id. 
+ * Returns USDC address based on chain Id.
  * Useful when you need to convert amount from a given token
  * to USDC (stable coin)
- * 
+ *
  * @param chainId ChainId
  * @returns string
  */
@@ -55,7 +53,7 @@ export const getUSDCAddress = (chainId: ChainId) => {
 
     case ChainId.FANTOM_MAINNET:
       return "0x04068DA6C83AFCFA0e13ba15A6696662335D5B75";
-    
+
     case ChainId.MAINNET:
       return "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48";
 
@@ -67,9 +65,9 @@ export const getUSDCAddress = (chainId: ChainId) => {
 
 /**
  * Returns the chain name given the id
- * 
+ *
  * @param id string
- * @returns string 
+ * @returns string
  */
 export const getChainVerbose = (id: string) => {
   switch (id) {
@@ -199,7 +197,7 @@ export const fetchRoundMetadata = async (
 /**
  * Generic function which handles how response is sent
  * for any API implemented in this service
- * 
+ *
  * @param res Response
  * @param code number
  * @param message string
@@ -227,8 +225,8 @@ export const handleResponse = (
 
 /**
  * Util function to get chainName for coingecko API calls
- * 
- * @param chainId 
+ *
+ * @param chainId
  * @returns { string, boolean}
  */
 export const getChainName = (chainId: ChainId) => {
@@ -327,7 +325,7 @@ export async function denominateAs(
     const avgAsTokenPrice =
       (asTokenPrices.startPrice + asTokenPrices.endPrice) / 2;
     const convertedAmount = amount * (avgAsTokenPrice / avgTokenPrice);
-    
+
     return {
       isSuccess: true,
       amount: convertedAmount,
@@ -346,9 +344,9 @@ export async function denominateAs(
 /**
  * This is temporary util function to support backward
  * compatibility with older subgraph version
- * 
+ *
  * TODO: remove after re-indexing mainnet subgraph
- * 
+ *
  * @param strategyName string
  * @returns string
  */
@@ -359,91 +357,18 @@ export const getStrategyName = (strategyName: string) => {
   return strategyName;
 }
 
-/**
- * summarizeRound is an async function that summarizes a round of voting by counting the number of contributions, the number of unique contributors, the total amount of contributions in USD, and the average contribution in USD.
- *
- * @param {ChainId} chainId - The id of the chain to fetch token prices from.
- * @param {RoundMetadata} roundMetadata - An object containing metadata about the round, including the start and end times and the token being voted on.
- * @param {QFVote[]} contributions - An array of QFVote objects representing the contributions made in the round.
- * @return {Promise<QFVoteSummary>} - An object containing the summarized data for the round.
- */
-export const summarizeQFVotes = async (
-  chainId: ChainId,
-  roundMetadata: RoundMetadata,
-  contributions: QFVote[]
-): Promise<QFVoteSummary> => {
-  // Create an object to store the sums
-  const summary: QFVoteSummary = {
-    contributionCount: 0,
-    uniqueContributors: 0,
-    totalContributionsInUSD: "",
-    averageUSDContribution: "",
-  };
-
-  const summaryContributions: any = {
-    contributions: {},
-    contributors: [],
-  };
-
-  // Iterate over the array of objects
-  contributions.forEach((item: QFVote) => {
-    // Get the token
-    const token = item.token;
-    const contributor = item.contributor;
-
-    // Initialize the sum for the token if it doesn't exist
-    if (!summaryContributions.contributions[token]) {
-      summaryContributions.contributions[token] = BigNumber.from("0");
-    }
-
-    // Initialize the contributor if it doesn't exist
-    if (!summaryContributions.contributors.includes(contributor)) {
-      summaryContributions.contributors.push(contributor);
-    }
-    // Update the sum for the token
-    summaryContributions.contributions[token] =
-      summaryContributions.contributions[token].add(item.amount);
-  });
-
-  let totalContributionsInUSD = 0;
-
-  const prices = await fetchTokenPrices(
-    chainId,
-    Object.keys(summaryContributions.contributions)
-  );
-
-  Object.keys(summaryContributions.contributions).map(async (tokenAddress) => {
-    const tokenAmount: number = Number(
-      formatUnits(summaryContributions.contributions[tokenAddress])
-    );
-
-    const conversionRate = prices[tokenAddress]?.usd;
-
-    const amountInUSD = tokenAmount * conversionRate;
-    totalContributionsInUSD += amountInUSD ? amountInUSD : 0;
-
-    return;
-  });
-
-  summary.totalContributionsInUSD = totalContributionsInUSD.toString();
-  summary.contributionCount = contributions.length;
-  summary.uniqueContributors = summaryContributions.contributors.length;
-  summary.averageUSDContribution = (
-    Number(summary.totalContributionsInUSD) / summary.uniqueContributors
-  ).toString();
-
-  return summary;
-};
 
 /**
- * fetchTokenPrices is an async function that retrieves the current prices of the tokens in tokenAddresses in USD.
- * If the native token of the chain with id chainId is included in tokenAddresses, its price is also included in the returned data.
+ * fetchTokenPrices is an async function that retrieves the current prices
+ * of the tokens in tokenAddresses in USD.
+ * If the native token of the chain with id chainId is included in
+ * tokenAddresses, its price is also included in the returned data.
  *
  * @param {ChainId} chainId - The id of the chain to retrieve the native token's price from.
  * @param {string[]} tokenAddresses - The addresses of the tokens to retrieve prices for.
  * @return {Promise<any>} - An object containing the token addresses as keys and their prices in USD as values.
  */
-const fetchTokenPrices = async (chainId: ChainId, tokenAddresses: string[]) => {
+export const fetchTokenPrices = async (chainId: ChainId, tokenAddresses: string[]) => {
   let data: any = {};
   try {
     const { chainName } = getChainName(chainId);
